backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.62
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.3.0
PyPDF2==3.0.1
python-multipart==0.0.6
gunicorn==21.2.0